only add the htlc pair to .htlcForwards after the receiving peer has committed the state.
this prevents the weird issue in which we add because all looks right and we send the state_update
to the receiving peer, but they just ignore us and reconnect again.

now the sender (another HC peer) can't get a resolution for the HTLC they've forwarded through us
because we are forever expecting the receiver to resolve -- and they can't resolve because it's not
pending for them.

the other alternative -- to check, for every incoming HTLC that is destined to another hosted peer,
if this other destined hosted peer has indeed the HTLC in their state -- is bug-prone, subjected to
data races that can cause loss of money for us, i.e.: between the moment we write .htlcForwards and
the moment we receive a confirmation for the updated state of the receiving peer the outgoing state
of this HTLC will be undetermined, and if the sender connects we will check and conclude there is no
outgoing HTLC, so fail the incoming one and return the balance to the initial sender -- in this case
we lose money.