make it so lcssNext is not a lazy val anymore.
this was causing the most horrific issues with we moving back in time because every time
we would try to use lcssNext it would have a cached value from a time ago and it was horrible.

the caching worked before our refactor from a week (?) ago because the state in which lcssNext
was based was tightly coupled to the actual valid lcss -- but now that we're using the actual
valid lcss as whatever we have on the database the `ChannelState` value can continue to be the
same even though the lcss has changed on the database. this means lcssNext may continue to be
the same even though we've moved onwards (on the database) on the actual lcss our peer that was
committed.

using a `def` instead of a `lazy val` should fix this.

but since a `def` is much more expensive (given the huge amount of times we use `lcssNext` in
some places) we've removed the signature calculation step from it, and moved that to the actual
call of `stateUpdate` (which didn't take any arguments before) and `stateOverride` (idem).

also took the opportunity to turn `.getPrivateKey()` from the NodeInterface in a
`lazy val privateKey` because why not? it was a super expensive operation anyway, and that for
sure is not expected to change.

(I think my previous rationale for not using a val or lazy val here was that I didn't want the
private key to be kept on memory, but I think that's not a valid concern.)