state_updates are sent with the latest uncommitted state always.
before we were sending one state_update for each htlc add/remove message, always corresponding
to whatever the state was immediately after adding/removing the htlc.

now we will send the latest state, which means that if we get a new htlc add/remove in between
the time we had sent the first htlc add/remove and the time we're about to send the state_update,
we will send the state_update already containing the two new htlc changes.

I think this will fix the errors related to these race conditions that happen on the immortan side.

to make things better we're also using a `lazy val` to ensure that each state_update is only sent
once -- even if it is the result of two htlc changes happening sequentially (without this we would
send two equal state_updates in this case, which wouldn't be harmful I think, but it would be uglier.)

---

if it doesn't work we'll have to resort to the cats-effect fix idea of making everything synchronous,
 discussed and explained at https://discord.com/channels/632150470000902164/635668881951686686/995350848735563877