more specific failure errors, and move the fee checking to the incoming channel.
rather than applying it on the outgoing channel, which was wrong, right?

in A-->B-->C the fee is paid from A to B according to the AB channel fee policy.
therefore if we're getting an `htlc_accepted` from lightningd it must have already checked
the incoming fee for that so fine -- we remove the fee checking from that.

then we check the fee when we're getting an HTLC from the hosted peer instead.